Fix update event layout Fix bugs in PDF export Fix validation

+++ public/admin_page/js/post_list.js
@@ -179,23 +179,26 @@ function showPosts(data) {
 }
 
 function deletePost(event) {
-    event.preventDefault();
+    var confirmation = confirm($LISTPOST_CONFIRM_DELETE);
+    if(confirmation == true) {
+        event.preventDefault();
 
-    // If they did, do our delete
-    $.ajax({
-        type: 'DELETE',
-        url: '/posts/deletepost/' + $(this).attr('rel')
-    }).done(function( response ) {
+        // If they did, do our delete
+        $.ajax({
+            type: 'DELETE',
+            url: '/posts/deletepost/' + $(this).attr('rel')
+        }).done(function( response ) {
 
-        // Check for a successful (blank) response
-        if (response.msg === '') {
-        }
-        else {
-            alert('Error: ' + response.msg);
-        }
+            // Check for a successful (blank) response
+            if (response.msg === '') {
+            }
+            else {
+                alert('Error: ' + response.msg);
+            }
 
-        // Update the table
-        populateTables();
+            // Update the table
+            populateTables();
 
-    });
+        });
+    }    
 }
